+++ assets/worldspaces/test.lua
@@ -1,17 +1,14 @@
 updateCount = 0
 
-lastCamEty = nil
-nextCamEty = 2
-
 function onUpdate(ety)
   updateCount = updateCount + 1
   
   if updateCount > 120 then
     updateCount = 0
-    local camEty = getCameraEntity()
-    setCameraEntity(nextCamEty)
-    lastCamEty = nextCamEty
-    nextCamEty = camEty
+    
+    state = getGlobalVar("orb-toggle-state") or false
+  
+    setGlobalBool("orb-toggle-state", not state)
   end
 end