+++ library/Kathu/Entity/System.hs
@@ -28,6 +28,7 @@ import           Kathu.Entity.Physics.Floor (FloorPropEntity)
 import           Kathu.Entity.Time
 import           Kathu.Graphics.Palette (PaletteManager, staticManager)
 import           Kathu.Scripting.Variables (Variables)
+import           Kathu.Util.Apecs
 import           Kathu.Util.Types (IDMap)
 import           Kathu.World.Stasis
 import           Kathu.World.Tile (Tile(..), TileID(..), tileID, TileState(..), tileTextID, unTileID)
@@ -76,6 +77,12 @@ instance Semigroup Variables where (<>) = mappend
 instance Monoid Variables where mempty = error "Attempted to access Variables global component before it has been initialized"
 instance Component Variables where type Storage Variables = Global Variables
 
+-- | A generic counter that will increment to produce a unique number
+newtype  Counter = Counter {unCounter :: Int}
+instance Semigroup Counter where (<>) = mappend
+instance Monoid Counter where mempty  = Counter 0
+instance Component Counter where type Storage Counter = Global Counter
+
 -- Entity functions
 
 stepLogicTime :: forall w m. (Has w m LogicTime, MonadIO m) => Word32 -> SystemT w m ()
@@ -86,6 +93,12 @@ stepRenderTime !dT = modify global $ \(RenderTime t) -> RenderTime (t + dT)
 
 stepWorldTime :: forall w m. (Has w m WorldTime, MonadIO m) => Word32 -> SystemT w m ()
 stepWorldTime !dT = modify global $ \(WorldTime t) -> WorldTime (t + fromIntegral dT)
+
+getNextFromCounter :: forall w m. (ReadWrite w m Counter, MonadIO m) => SystemT w m Int
+getNextFromCounter = do
+    Counter i <- get global
+    global    $= Counter (i + 1)
+    return i
 
 -------------
 -- Unsafe! --