+++ library/Kathu/Scripting/Event.hs
@@ -1,11 +1,14 @@
 {-# LANGUAGE OverloadedStrings #-}
 
 module Kathu.Scripting.Event
-    ( ScriptEvent(..)
+    ( ScriptEvent
     , EventFlag
     , noEventFlags
     , mkEventFlags
     , isEventSet
+    , onUpdate
+    , onInit
+    , onDestroy
     ) where
 
 import           Data.Aeson
@@ -13,10 +16,9 @@ import           Data.Bits
 import           Data.Word
 import qualified Data.Foldable    as F
 
-data ScriptEvent
-    = OnUpdate
+newtype ScriptEvent = ScriptEvent Word32
     
-data EventFlag = EventFlag Word32
+newtype EventFlag = EventFlag Word32
 
 noEventFlags :: EventFlag
 noEventFlags = EventFlag 0
@@ -25,17 +27,22 @@ mkEventFlags :: Foldable t => t ScriptEvent -> EventFlag
 mkEventFlags = F.foldl' (flip enableEventFlag) noEventFlags
 
 enableEventFlag :: ScriptEvent -> EventFlag -> EventFlag
-enableEventFlag e (EventFlag f) = EventFlag . (f .|.) $ case e of
-    OnUpdate -> 1
+enableEventFlag (ScriptEvent e) (EventFlag f) = EventFlag $ f .|. e
 
 isEventSet :: ScriptEvent -> EventFlag -> Bool
-isEventSet e (EventFlag f) = 0 /= case e of
-    OnUpdate -> 1 .&. f
+isEventSet (ScriptEvent e) (EventFlag f) = 0 /= e .&. f
+
+onUpdate, onInit, onDestroy :: ScriptEvent
+onUpdate  = ScriptEvent $ 2 ^ (0 :: Int)
+onInit    = ScriptEvent $ 2 ^ (1 :: Int)
+onDestroy = ScriptEvent $ 2 ^ (2 :: Int)
 
 instance FromJSON ScriptEvent where
     parseJSON = withText "ScriptEvent" $ \txt -> case txt of
-        "on-update" -> pure OnUpdate
-        e           -> fail $ "Unkown ScriptEvent " ++ show e
+        "on-update"  -> pure onUpdate
+        "on-init"    -> pure onInit
+        "on-destroy" -> pure onDestroy
+        e            -> fail $ "Unkown ScriptEvent " ++ show e
 
 instance FromJSON EventFlag where
     parseJSON = withArray "EventFlag" $ \arr -> mkEventFlags <$> sequence (parseJSON <$> arr)