+++ library/Kathu/Scripting/Lua/Component.hs
@@ -2,28 +2,33 @@
 {-# LANGUAGE DataKinds        #-}
 {-# LANGUAGE ExplicitForAll   #-}
 {-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 
 module Kathu.Scripting.Lua.Component (registerComponentFunctions) where
 
 import           Apecs
 import           Apecs.Physics
-import           Data.Text               (Text)
-import qualified Data.Vector             as Vec
+import           Data.Text                         (Text)
+import qualified Data.Vector                       as Vec
 import           Foreign.Lua
-import           Linear.V2               (V2(..))
+import           Linear.V2                         (V2(..))
 
 import           Kathu.Entity.Components
+import           Kathu.Graphics.Drawable           (Render(..), RenderSprite(..), switchAnimationByID)
+import           Kathu.Scripting.ExternalFunctions
 import           Kathu.Util.Apecs
-import           Kathu.Util.Types        (unID)
+import           Kathu.Util.Types                  (mkIdentifier, unID)
 
-registerComponentFunctions :: forall w. (Has w IO Physics, ReadWriteEach w IO '[Force, Identity, Mass, MovingSpeed, Position, Tags, Velocity]) => w -> Lua ()
-registerComponentFunctions world = do
+-- ExternalFunctions is unused in this, but is included here since it might be in the future, mirrors the Global's function signature, and acts as a Proxy for g
+registerComponentFunctions :: forall w g. (Has w IO Physics, Members w IO (Render g), ReadWriteEach w IO '[Force, Identity, Mass, MovingSpeed, Position, Render g, Tags, Velocity]) => w -> ExternalFunctions w g -> Lua ()
+registerComponentFunctions world _ = do
     registerHaskellFunction "getIdentifier"  (getIdentifier world)
     registerHaskellFunction "getName"        (getName world)
     registerHaskellFunction "getDescription" (getDescription world)
     registerHaskellFunction "getTags"        (getTags world)
     registerHaskellFunction "getMovingSpeed" (getMovingSpeed world)
     registerHaskellFunction "setMovingSpeed" (setMovingSpeed world)
+    registerHaskellFunction "setAnimation"   (setAnimation (Proxy :: Proxy g) world)
     registerHaskellFunction "getMass"        (getMass world)
     registerHaskellFunction "setMass"        (setMass world)
     registerHaskellFunction "getPosition"    (getPosition world)
@@ -71,6 +76,21 @@ getMovingSpeed !world !etyID = liftIO . Apecs.runWith world $ do
 setMovingSpeed :: forall w. (ReadWrite w IO MovingSpeed) => w -> Int -> Double -> Lua ()
 setMovingSpeed !world !etyID s = liftIO . Apecs.runWith world $ do
     (Entity etyID) $= MovingSpeed s
+
+--------------
+-- Graphics --
+--------------
+
+setAnimation :: forall w g. (Members w IO (Render g), ReadWrite w IO (Render g)) => Proxy g -> w -> Int -> Text -> Lua ()
+setAnimation _ !world !etyID !animID = liftIO . Apecs.runWith world $ do
+    let ety = Entity etyID
+        changeAnim (RSAnimated anim) = RSAnimated $ switchAnimationByID (mkIdentifier animID) anim
+        changeAnim e                 = e
+
+    mrender :: Maybe (Render g) <- getIfExists ety
+    case mrender of
+        Nothing     -> return ()
+        Just render -> ety $= Render (changeAnim <$> unRender render)
 
 -------------
 -- Physics --